fix(ffi): 修正 wimlib_get_version_string 绑定为 UTF-16(tchar)
GetVersionString 此前声明为 `-> *const u8`，version_string() 也用
from_utf8_lossy 按字节解码。但 wimlib_get_version_string 返回
`const wimlib_tchar *`，在 Windows 上即 wchar_t(UTF-16LE)——按 u8 读
会在第二个字节(0x00)处截断，"1.14.4" 只剩 "1"。

- 绑定返回类型 *const u8 -> *const u16
- version_string() 改用统一的 util::from_wide 解码，移除内联手写的宽字符扫描
  (符合 CLAUDE.md 第 8 节"宽字符转换集中在一个 helper")
- info 命令打印 wimlib 库版本，使该函数真正被调用(此前是死代码)

+++ src/ffi/types.rs
@@ -119,7 +119,9 @@ pub type GlobalInit = unsafe extern "C" fn(init_flags: c_int) -> c_int;
 pub type GlobalCleanup = unsafe extern "C" fn();
 pub type Free = unsafe extern "C" fn(wim: *mut WimStruct);
 pub type GetErrorString = unsafe extern "C" fn(code: c_int) -> *const u16;
-pub type GetVersionString = unsafe extern "C" fn() -> *const u8;
+// wimlib_get_version_string 返回 `const wimlib_tchar *`，在 Windows 上即
+// wchar_t（2 字节 UTF-16LE），因此是 *const u16 而非 *const u8。
+pub type GetVersionString = unsafe extern "C" fn() -> *const u16;
 pub type OpenWimWithProgress = unsafe extern "C" fn(
     wim_file: *const u16,
     open_flags: c_int,
